requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0

pandas==2.2.2
openpyxl==3.1.2


gspread==5.10.0
google-auth==2.29.0
schedule==1.2.1

